requirements.txt
Django == 4.2
djangorestframework == 3.15.2
drf-yasg == 1.21.7
